--- runtime/kotlin/src/main/resources/META-INF/services/javax.script.ScriptEngineFactory
@@ -1,1 +1,0 @@
-org.jetbrains.kotlin.script.jsr223.KotlinJsr223JvmLocalScriptEngineFactory
