--- runtime/kotlin/src/main/resources/META-INF/services/org.apache.camel.k.RoutesLoader
@@ -1,1 +1,0 @@
-org.apache.camel.k.kotlin.KotlinRoutesLoader
